requirements.txt
selenium==4.1.3
python-dotenv==1.0.1